[dl, rebase] SR830 semi-working but has read errors.
I can talk to the
lock in and change at least the SENS parameter that sets the
sensitivity.  there is a bug that forces you to quarry the SENS
twice!!!! before it will return the correct value.

davidlow: added NIdaq first draft based on brian and david's SQUID_IV.
Not tested, not completed.  The majority of NIdaq has been written.
git sort of incorporated.